Refactor beads-core state module
Split `crates/beads-core/src/state.rs` into submodules:
- `canonical.rs`: `CanonicalState` and core logic.
- `labels.rs`: `LabelStore` and `LabelState`.
- `deps.rs`: `DepStore` and `DepIndexes`.
- `notes.rs`: `NoteStore`.

Re-export types in `mod.rs` to maintain API compatibility.
Update imports and visibility as needed.
